+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Download rather than stream audio.
 - Start playing audio from current verse.
 - Play verse when verse number tapped (if audio player is visible).
+- Fix navigation when in repeat verse audio mode.
+- Fix live swapping audio voices.
 
 ## 0.14.0 - January 10, 2025